The main project is RateMyProfessor.py, everything else so far is practice.
The purpose of this project is to make a crucial part of every college
    students life, class selection, a bit easier. By allowing students
    to search all the professors they want in one go, instead
    of searching each professor individually, the amount of stress in
    picking classes will hopefully be quite less, and certainly more
    efficient.

Synopsis of RateMyProfessor:
    The User inputs the names of each professor they want to search. The
    program will use webscraping to find the stats on each professor
    and return them as a list. The stats found are namely:
        average rating, average difficulty, would take again percentage,
        number of reviews, and a link to the professors page on
        RateMyProfessor.

So far this project has taken me probably close to 30 hours, but
most of that was learning the various aspects of python -since this is
my first real project in python- as well as lots of trial and error and
revisions of the processes for going from input to reaching an output.

- This project uses the following packages: requests, json, and
    BeautifulSoup4.

- The website that is scraped is https://www.ratemyprofessors.com/

Shout out to John Watson Rooney and his youtube channel, which was a
great help for me, and tought me a lot about web scraping.
https://www.youtube.com/@JohnWatsonRooney

Note: This program currently only works for Auburn University,
but I plan on adding functionality to work with every University
on the RateMyProfessor website.

P.S: I really didn't use the git repository responsibly at all for
    this project, should've started it from the get-go. This is a winter
    break project that has been a brain blistering blast so far,
    and something I am proud of, save my lack of forethought on the git
    end of things.